Implemented Playlist Controller
