+++ src/main/java/com/galvanize/musicshare/controller/PlaylistController.java
@@ -1,0 +1,16 @@
+package com.galvanize.musicshare.controller;
+
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("api/v1/playlist")
+
+public class PlaylistController {
+
+    @PostMapping
+    public void createPlaylist(){
+
+    }
+}
