+++ src/main/java/com/galvanize/musicshare/entity/Song.java
@@ -1,0 +1,17 @@
+package com.galvanize.musicshare.entity;
+
+import lombok.Data;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+
+@Entity
+@Data
+public class Song {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long songId;
+    private String songName;
+}
