--- src/main/java/entity/PlaylistSongMapping.java
@@ -1,17 +1,0 @@
-package entity;
-
-import lombok.Data;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-
-@Entity
-@Data
-public class PlaylistSongMapping {
-    @Id
-    @GeneratedValue
-    private Long id;
-    private String playlistName;
-    private String songName;
-}
