+++ src/test/java/com/galvanize/musicshare/controller/PlaylistControllerTest.java
@@ -1,0 +1,25 @@
+package com.galvanize.musicshare.controller;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.web.servlet.MockMvc;
+
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+
+@SpringBootTest
+@AutoConfigureMockMvc
+public class PlaylistControllerTest {
+
+    @Autowired
+    private MockMvc mockMvc;
+
+    @Test
+    public void createPlaylist() throws Exception {
+        mockMvc.perform(
+                post("/api/v1/playlist"))
+                .andExpect(status().isOk());
+    }
+}
